+++ src/modules/arrowhead/common/JSON/GetArrayResponseFromJSON_fbt.cpp
@@ -10,9 +10,9 @@
  *   Jose Cabral - initial implementation
  *******************************************************************************/
 
-#include "GetArrayResponseFromJSON.h"
+#include "GetArrayResponseFromJSON_fbt.h"
 #ifdef FORTE_ENABLE_GENERATED_SOURCE_CPP
-#include "GetArrayResponseFromJSON_gen.cpp"
+#include "GetArrayResponseFromJSON_fbt_gen.cpp"
 #endif
 
 #include "ArrowheadJSONHelper.h"
