+++ src/modules/arrowhead/eventHandler/ArrowheadPublishAdp_atp.cpp
@@ -10,9 +10,9 @@
  *   Jose Cabral - initial implementation
  *******************************************************************************/
 
-#include "ArrowheadPublishAdp.h"
+#include "ArrowheadPublishAdp_atp.h"
 #ifdef FORTE_ENABLE_GENERATED_SOURCE_CPP
-#include "ArrowheadPublishAdp_gen.cpp"
+#include "ArrowheadPublishAdp_atp_gen.cpp"
 #endif
 
 DEFINE_ADAPTER_TYPE(FORTE_ArrowheadPublishAdp, g_nStringIdArrowheadPublishAdp)
