+++ src/modules/arrowhead/eventHandler/HTTP/SubscribeEventHTTP.h
@@ -15,7 +15,7 @@
 
 #include <cfb.h>
 #include <typelib.h>
-#include "SubscribeEventAdp.h"
+#include "SubscribeEventAdp_atp.h"
 
 class FORTE_SubscribeEventHTTP: public CCompositeFB{
   DECLARE_FIRMWARE_FB(FORTE_SubscribeEventHTTP)
