+++ src/modules/arrowhead/eventHandler/SubscribeEventAdp_atp.cpp
@@ -10,9 +10,9 @@
  *   Jose Cabral - initial implementation
  *******************************************************************************/
 
-#include "SubscribeEventAdp.h"
+#include "SubscribeEventAdp_atp.h"
 #ifdef FORTE_ENABLE_GENERATED_SOURCE_CPP
-#include "SubscribeEventAdp_gen.cpp"
+#include "SubscribeEventAdp_atp_gen.cpp"
 #endif
 
 DEFINE_ADAPTER_TYPE(FORTE_SubscribeEventAdp, g_nStringIdSubscribeEventAdp)
