+++ src/modules/arrowhead/orchestrator/CMakeLists.txt
@@ -20,7 +20,7 @@ forte_add_include_directories(${CMAKE_CURRENT_SOURCE_DIR})
 forte_add_sourcefile_hcpp(forte_orchestrationform forte_preferredprovider forte_servicerequestform)
 
 #Orchestrator
-forte_add_sourcefile_hcpp(FieldsToPreferredProvider FieldsToServiceRequestForm OrchestratorRequestAdp RequestOrchestrationForm OrchestrationForm2OrchestrationForm)
+forte_add_sourcefile_hcpp(FieldsToPreferredProvider FieldsToServiceRequestForm OrchestratorRequestAdp_atp RequestOrchestrationForm OrchestrationForm2OrchestrationForm_fbt)
 
 if(FORTE_COM_HTTP)
   add_subdirectory(HTTP)
