+++ src/modules/arrowhead/orchestrator/RequestOrchestrationForm.h
@@ -19,7 +19,7 @@
 #include <forte_servicerequestform.h>
 #include <forte_wstring.h>
 #include <forte_array.h>
-#include "OrchestratorRequestAdp.h"
+#include "OrchestratorRequestAdp_atp.h"
 
 class FORTE_RequestOrchestrationForm: public CCompositeFB{
   DECLARE_FIRMWARE_FB(FORTE_RequestOrchestrationForm)
