+++ src/modules/arrowhead/serviceRegistry/HTTP/RegisterServiceHTTP.h
@@ -15,7 +15,7 @@
 
 #include <cfb.h>
 #include <typelib.h>
-#include "RegisterServiceAdp.h"
+#include "RegisterServiceAdp_atp.h"
 
 class FORTE_RegisterServiceHTTP: public CCompositeFB{
   DECLARE_FIRMWARE_FB(FORTE_RegisterServiceHTTP)
