+++ src/modules/arrowhead/serviceRegistry/QueryServicesAdp_atp.h
