+++ src/modules/arrowhead/serviceRegistry/ServiceRegistryEntry2ServiceRegistryEntry_fbt.h
